рефакторинг и calcDiffBtwAverageAndMedianPriceBtwVvoAndTlv

+++ src/main/java/com/idea/platform/Service/TicketFilter.java
@@ -6,11 +6,8 @@ import java.util.List;
 import java.util.Map;
 
 public interface TicketFilter {
-    // Минимальное время полета между городами Владивосток и Тель-Авив для каждого авиаперевозчика
     public Map<String, Long> calcMinTimeBtwVvoAndTlv(List<Ticket> tickets);
 
-
-    // Разницу между средней ценой  и медианой для полета между городами  Владивосток и Тель-Авив
-    public Integer calcDiffBtwAverageAndMedianPriceBtwVvoAndTlv(List<Ticket> tickets);
+    public long calcDiffBtwAverageAndMedianPriceBtwVvoAndTlv(List<Ticket> tickets);
 
 }
